requirements.txt
amqp==2.2.2
anyjson==0.3.3
billiard==3.5.0.3
chardet==2.2.1
Django==1.11.10
django-background-tasks==1.1.11
django-compat==1.0.14
django-cors-headers==2.1.0
django-crontab==0.7.1
djangorestframework==3.7.0
djangorestframework-jwt==1.11.0
docopt==0.6.1
docutils==0.7
egenix-mx-base==3.1.3
iniparse==0.4
iotop==0.6
IPy==0.75
javapackages==1.0.0
kitchen==1.1.1
kombu==4.1.0
lxml==3.2.1
MySQL-python==1.2.3
netaddr==0.7.5
netifaces==0.10.4
ngxtop==0.0.2
numpy==1.5.0
olefile==0.44
PIL==1.1.7
Pillow==4.3.0
policycoreutils-default-encoding==0.1
psycopg2==2.5.4
pycrypto==2.3
pycurl==7.19.0
pygpgme==0.3
PyJWT==1.5.3
pyliblzma==0.5.3
pyparsing==1.5.6
pytz==2017.3
pyxattr==0.5.1
reportlab==2.4
seobject==0.1
sepolicy==1.1
six==1.11.0
tabulate==0.7.2
urlgrabber==3.10
vine==1.1.4
yum-metadata-parser==1.1.4
